get-tif analysis action, missing link handling,
